refactor: integrate SampleForm and TestcaseForm
